requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
docstring-parser==0.16
    # via google-cloud-aiplatform
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-api-core==2.24.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-cloud-aiplatform==1.84.0
    # via vertex-chat-permission (pyproject.toml)
google-cloud-bigquery==3.30.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via google-cloud-aiplatform
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.69.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-resource-manager
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
narwhals==1.31.0
    # via altair
numpy==2.2.4
    # via
    #   pandas
    #   pydeck
    #   shapely
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.1.0
    # via streamlit
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==5.29.3
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
    #   streamlit
psycopg2==2.9.10
    # via vertex-chat-permission (pyproject.toml)
pyarrow==19.0.1
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via google-cloud-aiplatform
pydantic-core==2.27.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2025.1
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   streamlit
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
ruff==0.11.0
    # via vertex-chat-permission (pyproject.toml)
shapely==2.0.7
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sqlalchemy==2.0.39
    # via vertex-chat-permission (pyproject.toml)
streamlit==1.43.2
    # via vertex-chat-permission (pyproject.toml)
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   google-cloud-aiplatform
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
